GetTextureParam.cs
