+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/GetTextureParam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f1dc05e4970e724e9f50c0a5a67a25e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
